UI: Move VolumeSlider widget to volume control
